requirements.txt
astroid==2.4.2
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.1
mccabe==0.6.1
pylint==2.5.3
python-docx==0.8.10
six==1.15.0
toml==0.10.1
wrapt==1.12.1
